IGNITE-7993 Striped pool can't be disabled
* Renamed StripedExecutor to StripedExecutorImpl
* Added StripedExecutor interface
* Added StripedExecutorProxy (extends IgniteThreadPoolExecutor impl StripedExecutor)
* Added performance optimization suggestion
* Added StripedExecutorProxyTest
