+++ modules/core/src/main/java/org/apache/ignite/internal/IgniteKernal.java
@@ -1470,6 +1470,11 @@ public class IgniteKernal implements IgniteEx, IgniteMXBean, Externalizable {
 
         if (BinaryMarshaller.available() && (cfg.getMarshaller() != null && !(cfg.getMarshaller() instanceof BinaryMarshaller)))
             perf.add("Use default binary marshaller (do not set 'marshaller' explicitly)");
+
+        if (IgniteSystemProperties.getBoolean(IgniteSystemProperties.IGNITE_STRIPED_POOL_DISABLED)) {
+            perf.add("Enable striped pool (remove -DIGNITE_STRIPED_POOL_DISABLED=true" +
+                " or set -DIGNITE_STRIPED_POOL_DISABLED=false)");
+        }
     }
 
     /**
